+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
+- refectory interfaces in core
+- add work for manager cronjobs
+- add cronjob
+- add test for cron job list user
 
 ## [1.1.0]