+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add test for cron job list user
 - add cache structure
 - refectory test unit list user
-- Added unit testing in the cache at the infra layer 
+- Added unit testing in the cache at the infra layer
+- Fix cache in use case list users
 
 ## [1.1.0]