+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
+
+## [2.0.0]
 - refectory interfaces in core
 - add work for manager cronjobs
 - add cronjob